marketplace/requirements.txt
flask ~= 1.1
grpcio-tools ~= 1.30
Jinja2
pytest ~= 5.4